refactor: react-query와 api요청 분리
